Convert into test to check _determine_api_scope(), rather than of expected outcomes across a variety of libraries.
